Issue #377 - use the VMUSB drivers in place rather than as copies
Import MDPP 16 channel QDC support from the vmusb
