+++ main/mvlc/Makefile.am
@@ -67,11 +67,12 @@ IMPORTED_MODULES=$(IMPORT_DIR)/CMQDC32RdoHdwr.h \
 	$(IMPORT_DIR)/CMADCChain.h $(IMPORT_DIR)/CMADCChain.cpp  \
 	$(IMPORT_DIR)/CMADCScaler.h $(IMPORT_DIR)/CMADCScaler.cpp \
 	$(IMPORT_DIR)/CMarker.h $(IMPORT_DIR)/CMarker.cpp \
-	$(IMPORT_DIR)/ CMDPP.h
+	$(IMPORT_DIR)/ CMDPP.h \
+	$(IMPORT_DIR)/CMDPP16QDC.cpp  $(IMPORT_DIR)/CMDPP16QDC.h
 
 LOCAL_MODULES= 	MADC32Registers.h \
 	CMTDC32.h CMTDC32.cpp \
-	CMDPP16QDC.cpp  CMDPP16QDC.h  CMDPP32QDC.cpp  CMDPP32QDC.h \
+	CMDPP32QDC.cpp  CMDPP32QDC.h \
 	CMDPP32SCP.cpp  CMDPP32SCP.h   CMDPPQDC.h \
 	CNADC2530.h CNADC2530.cpp CV977.h CV977.cpp \
 	CV1495sc.h CV1495sc.cpp CV1729.h CV1729.cpp \
